feat: add fertile soil element.

+++ src/Game/Elements/Liquids/Acid.cs
@@ -1,4 +1,5 @@
-﻿using StardustSandbox.Elements.Solids.Immovables;
+﻿using StardustSandbox.Enums.Elements;
+using StardustSandbox.Randomness;
 
 namespace StardustSandbox.Elements.Liquids
 {
@@ -13,20 +14,27 @@ namespace StardustSandbox.Elements.Liquids
                     continue;
                 }
 
-                switch (neighbors.GetSlotLayer(i, context.Layer).Element)
+                switch (neighbors.GetSlotLayer(i, context.Layer).Element.Index)
                 {
-                    case Acid:
-                    case Wall:
-                    case Clone:
-                    case Void:
+                    case ElementIndex.Acid:
+                    case ElementIndex.Wall:
+                    case ElementIndex.Clone:
+                    case ElementIndex.Void:
+                    case ElementIndex.DownwardPusher:
+                    case ElementIndex.UpwardPusher:
+                    case ElementIndex.LeftwardPusher:
+                    case ElementIndex.RightwardPusher:
                         continue;
 
                     default:
                         break;
                 }
 
-                context.DestroyElement();
-                context.DestroyElement(neighbors.GetSlot(i).Position, context.Layer);
+                if (SSRandom.GetBool())
+                {
+                    context.DestroyElement(neighbors.GetSlot(i).Position, context.Layer);
+                    context.DestroyElement();
+                }
             }
         }
     }
